requirements.txt
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
Flask-Cors==3.0.10
google-auth==2.16.2
google-auth-oauthlib==1.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
oauthlib==3.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==1.0.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
urllib3==1.26.14
Werkzeug==2.2.3

PyJWT~=2.6.0
pymongo~=4.3.3
PyYAML==6.0